+++ tests/test_widget_removing.py
@@ -4,24 +4,12 @@ from textual.widget import Widget
 from textual.widgets import Static, Button
 from textual.containers import Container
 
-async def await_remove_standin():
-    """Standin function for awaiting removal.
-
-    These tests are being written so that we can go on and make remove
-    awaitable, but it would be good to have some tests in place *before* we
-    make that change, but the tests need to await remove to be useful tests.
-    So to get around that bootstrap issue, we just use this function as a
-    standin until we can swap over.
-    """
-    await asyncio.sleep(0) # Until we can await remove.
-
 async def test_remove_single_widget():
     """It should be possible to the only widget on a screen."""
     async with App().run_test() as pilot:
         await pilot.app.mount(Static())
         assert len(pilot.app.screen.children) == 1
-        pilot.app.query_one(Static).remove()
-        await await_remove_standin()
+        await pilot.app.query_one(Static).remove()
         assert len(pilot.app.screen.children) == 0
 
 async def test_many_remove_all_widgets():
@@ -29,8 +17,7 @@ async def test_many_remove_all_widgets():
     async with App().run_test() as pilot:
         await pilot.app.mount(*[Static() for _ in range(1000)])
         assert len(pilot.app.screen.children) == 1000
-        pilot.app.query(Static).remove()
-        await await_remove_standin()
+        await pilot.app.query(Static).remove()
         assert len(pilot.app.screen.children) == 0
 
 async def test_many_remove_some_widgets():
@@ -38,8 +25,7 @@ async def test_many_remove_some_widgets():
     async with App().run_test() as pilot:
         await pilot.app.mount(*[Static(id=f"is-{n%2}") for n in range(1000)])
         assert len(pilot.app.screen.children) == 1000
-        pilot.app.query("#is-0").remove()
-        await await_remove_standin()
+        await pilot.app.query("#is-0").remove()
         assert len(pilot.app.screen.children) == 500
 
 async def test_remove_branch():
@@ -71,8 +57,7 @@ async def test_remove_branch():
             ),
         )
         assert len(pilot.app.screen.walk_children(with_self=False)) == 13
-        pilot.app.screen.children[0].remove()
-        await await_remove_standin()
+        await pilot.app.screen.children[0].remove()
         assert len(pilot.app.screen.walk_children(with_self=False)) == 7
 
 async def test_remove_overlap():
@@ -104,8 +89,7 @@ async def test_remove_overlap():
             ),
         )
         assert len(pilot.app.screen.walk_children(with_self=False)) == 13
-        pilot.app.query(Container).remove()
-        await await_remove_standin()
+        await pilot.app.query(Container).remove()
         assert len(pilot.app.screen.walk_children(with_self=False)) == 1
 
 async def test_remove_move_focus():
@@ -119,9 +103,44 @@ async def test_remove_move_focus():
         await pilot.press( "tab" )
         assert pilot.app.focused is not None
         assert pilot.app.focused == buttons[0]
-        pilot.app.screen.children[0].remove()
-        await await_remove_standin()
+        await pilot.app.screen.children[0].remove()
         assert len(pilot.app.screen.children) == 1
         assert len(pilot.app.screen.walk_children(with_self=False)) == 6
         assert pilot.app.focused is not None
         assert pilot.app.focused == buttons[9]
+
+async def test_widget_remove_order():
+    """A Widget.remove of a top-level widget should cause bottom-first removal."""
+
+    removals: list[str] = []
+
+    class Removable(Container):
+        def on_unmount( self, _ ):
+            removals.append(self.id if self.id is not None else "unknown")
+
+    async with App().run_test() as pilot:
+        await pilot.app.mount(
+            Removable(Removable(Removable(id="grandchild"), id="child"), id="parent")
+        )
+        assert len(pilot.app.screen.walk_children(with_self=False)) == 3
+        await pilot.app.screen.children[0].remove()
+        assert len(pilot.app.screen.walk_children(with_self=False)) == 0
+        assert removals == ["grandchild", "child", "parent"]
+
+async def test_query_remove_order():
+    """A DOMQuery.remove of a top-level widget should cause bottom-first removal."""
+
+    removals: list[str] = []
+
+    class Removable(Container):
+        def on_unmount( self, _ ):
+            removals.append(self.id if self.id is not None else "unknown")
+
+    async with App().run_test() as pilot:
+        await pilot.app.mount(
+            Removable(Removable(Removable(id="grandchild"), id="child"), id="parent")
+        )
+        assert len(pilot.app.screen.walk_children(with_self=False)) == 3
+        await pilot.app.query(Removable).remove()
+        assert len(pilot.app.screen.walk_children(with_self=False)) == 0
+        assert removals == ["grandchild", "child", "parent"]